Added some viewmodels
