+++ BluescreenSimulator/Models/Windows10Bluescreen.cs
@@ -2,7 +2,7 @@
 
 namespace BluescreenSimulator
 {
-    public class BluescreenData
+    public class Windows10Bluescreen : BluescreenBase
     {
         public string Emoticon { get; set; } = ":(";
 
@@ -17,16 +17,6 @@ namespace BluescreenSimulator
         public string SupportPerson { get; set; } = "If you call a support person, give them this info:";
 
         public string StopCode { get; set; } = "Stop code: DRIVER_IRQL_NOT_LESS_OR_EQUAL";
-
-        public string CmdCommand { get; set; } = null;
-
-        public Color BackgroundColor { get; set; } = Color.FromRgb(10, 112, 169);
-
-        public Color ForegroundColor { get; set; } = Colors.White;
-
-        public int Delay { get; set; } = 0;
-
-        public bool EnableUnsafe { get; set; } = false;
 
         public bool HideQR { get; set; } = false;
