+++ BluescreenSimulator/Views/BluescreenWindow9x.xaml
@@ -15,7 +15,7 @@ xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         Height="1080"
         Title="BluescreenWindow"
         Topmost="True"
-        d:DataContext="{d:DesignInstance viewModels:BluescreenDataViewModel, IsDesignTimeCreatable=True}">
+        d:DataContext="{d:DesignInstance viewModels:Windows10BluescreenViewModel, IsDesignTimeCreatable=True}">
 
     <Window.Style>
         <Style TargetType="Window">
